pyproject.toml
[build-system]
requires = ["setuptools>=67", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "relluna"
version = "0.2.0"
description = "Relluna - Infraestrutura de memória"
requires-python = ">=3.11"

dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.6.0",
    "python-dotenv>=1.0.1",
    "python-multipart>=0.0.9",
    "httpx>=0.27.0",
    "redis>=5.0.1",
    "boto3>=1.34.0",
    "pymongo>=4.8.0",
    "motor[srv]>=3.0.0",
    "pillow>=10.3.0",
    "pypdf>=5.0.0",
    "pytesseract>=0.3.10",
    "pymupdf>=1.23.0",
    "jinja2>=3.1.0",
    "jsonschema>=4.0.0",
    "numpy>=1.26.0",
    "opencv-python>=4.8.0",
    "azure-storage-blob>=12.19.0",
    "azure-identity>=1.15.0",
    "azure-keyvault-secrets>=4.7.0"
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "pytest-mock",
    "ruff>=0.6.9",
]
asr = [
    "openai-whisper>=20231117",
]

[tool.setuptools]
packages = ["relluna"]
